(Alex) Group Meeting & setting up MovieSelect
We had a meeting this evening with Blaine, Chance, and myself.  Tasks were divvied up with Blaine doing work on the database and Chance taking on group organization with the Trello page and the powerpoint for our next presentation.  I am beginning to weave together the HTML and Java for our project.